refactor: export bootstrap method for npm package
[AP-361]
